refactor: renamed users to user
